# backend/openapi/knowledge-platform-api.yaml
openapi: 3.1.0
info:
  title: RetailPartnerX Knowledge Platform API
  version: 0.1.0
  description: |
    Trust boundary for BFF ↔ Knowledge Platform.
    GraphRAG chat + multimodal label recognition (PDF/PNG/...).
servers:
  - url: http://localhost:8080
paths:
  /health:
    get:
      summary: Health
      responses:
        "200":
          description: OK
  /v1/chat:
    post:
      summary: GraphRAG chat (SSE when stream=true)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: ChatResponse or text/event-stream
  /v1/vision/label:
    post:
      summary: Recognize label from PDF/PNG/JPEG/WEBP/TIFF
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                page_limit:
                  type: integer
                  default: 3
      responses:
        "200":
          description: LabelResponse
  /v1/assets/{asset_id}/presign:
    get:
      summary: Presigned URL for raw/recognized object in MinIO/S3
      parameters:
        - in: path
          name: asset_id
          required: true
          schema:
            type: string
        - in: query
          name: kind
          schema:
            type: string
            enum: [raw, recognized, normalized]
      responses:
        "200":
          description: PresignResponse
  /v1/jobs:
    post:
      summary: Async job (202)
      responses:
        "202":
          description: Accepted
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Demo tokens — guest|associate|manager|compliance
  schemas:
    ChatRequest:
      type: object
      required: [message]
      properties:
        message:
          type: string
        session_id:
          type: string
        stream:
          type: boolean
          default: true